Dump a function with Element def included.
