+++ src/lib.rs
@@ -8,3 +8,4 @@
 
 pub mod error;
 pub mod parser;
+pub mod element;
