+++ src/parser.rs
@@ -4,80 +4,14 @@
 //
 
 use std::cell::Cell;
-use std::collections::HashMap;
 use std::fs::File;
 use std::io::prelude::*;
-use std::fmt;
+use std::path::Path;
+use std::io::Error;
+use std::io::ErrorKind;
 
+use super::element::*;
 use super::error::*;
-
-/// An individual element in an ABNF rule.
-#[derive(PartialEq, Clone)]
-pub enum Element {
-    /// rulename.
-    Rulename(String),
-    /// case insensitive string.
-    IString(String),
-    /// case seisitve string.
-    SString(String),
-    /// num-val.
-    NumberValue(u32),
-    /// range of num-val.
-    ValueRange((u32, u32)),
-    /// sequence of num-val.
-    ValueSequence(Vec<u32>),
-    /// prose-val.
-    ProseValue(String),
-    /// concatination.
-    Sequence(Vec<Repetition>),
-    /// alternation.
-    Selection(Vec<Repetition>),
-}
-
-impl fmt::Debug for Element {
-    fn fmt (&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match &*self {
-            Element::Rulename(s) => write!(f, "Element::Rulename({:?}.to_string())", s),
-            Element::IString(s) => write!(f, "Element::IString({:?}.to_string())", s),
-            Element::SString(s) => write!(f, "Element::SString({:?}.to_string())", s),
-            Element::NumberValue(n) => write!(f, "Element::NumberValue({:?})", n),
-            Element::ValueRange(t) => write!(f, "Element::ValueRange({:?})", t),
-            Element::ValueSequence(v) => write!(f, "Element::ValueSequence(vec!{:?})", v),
-            Element::ProseValue(s) => write!(f, "Element::ProseValue({:?}.to_string())", s),
-            Element::Sequence(v) => write!(f, "Element::Sequence(vec!{:?})", v),
-            Element::Selection(v) => write!(f, "Element::Selection(vec!{:?})", v),
-        }
-    }
-}
-
-/// Repeat.
-#[derive(PartialEq, Debug, Clone)]
-pub struct Repeat {
-    min: Option<usize>,
-    max: Option<usize>,
-}
-
-impl Repeat {
-    pub fn new(min: Option<usize>, max: Option<usize>) -> Repeat {
-        Repeat { min, max }
-    }
-}
-
-/// Element with repeat.
-#[derive(PartialEq, Debug, Clone)]
-pub struct Repetition {
-    repeat: Option<Repeat>,
-    element: Element,
-}
-
-impl Repetition {
-    pub fn new(repeat: Option<Repeat>, element: Element) -> Repetition {
-        Repetition { repeat, element }
-    }
-}
-
-/// Rulelist.
-type Rulelist = HashMap<String, Repetition>;
 
 /// ABNF Token type.
 #[derive(PartialEq, Debug)]
@@ -532,15 +466,16 @@ fn is_rule_delimiter(input: &str) -> bool {
 pub fn parse_file(filename: &str) -> std::io::Result<()> {
     let mut f = File::open(filename)?;
     let mut s = String::new();
+    let p = Path::new(filename).file_stem().ok_or(Error::new(ErrorKind::Other, "Invalid filename"))?;
+    let n1 = p.to_str().ok_or(Error::new(ErrorKind::Other, "Invalid filename"))?;
+    let n2 = str::replace(n1, ".", "_");
 
     f.read_to_string(&mut s)?;
     let mut parser = Parser::new(s);
 
     match parser.parse() {
         Ok(rl) => {
-//            println!("{:?}", rl);
-
-            rulelist_dump("hoge", &rl);
+            rulelist_dump(&n2, &rl)?;
         }
         Err(err) => {
             println!(
@@ -555,7 +490,12 @@ pub fn parse_file(filename: &str) -> std::io::Result<()> {
     Ok(())
 }
 
-pub fn rulelist_dump(name: &str, rl: &Rulelist) {
+pub fn rulelist_dump(name: &str, rl: &Rulelist) -> std::io::Result<()> {
+    let mut file = File::open("src/element.rs")?;
+    let mut elem = String::new();
+    file.read_to_string(&mut elem)?;
+    println!("{}", elem);
+
     println!(r#"pub fn get_{}_rulelist() -> Rulelist {{"#, name);
     println!(r#"    let mut rl: Rulelist = Rulelist::new();"#);
     println!(r#""#);
@@ -567,6 +507,8 @@ pub fn rulelist_dump(name: &str, rl: &Rulelist) {
     println!(r#""#);
     println!(r#"    rl"#);
     println!(r#"}}"#);
+
+    Ok(())
 }
 
 #[cfg(test)]
